Variable Speed Refactor and Additions.
This patch implements a more systematic way for new functions to be
added which allow for a hook into attacks which can change the speed of
a given attack if it's pre-conditions are met. This is done by creating
a new interface which any variable speed modifier can implement. This
allows all the expressiveness that comes with being a full java
function, and therefore easily can encode the trivial modifiers which
were previously "hacks" inside the `getWeaponSpeed` function these are
new self contained classes which implement the new interface.

The next step after re-implementing the existing variable speed
modifiers was to add new ones.

* [moved] Blood Moon
* [moved] Rapid/Ranged Salamander
* [moved] Leagues
* [moved] Red Keris Special Attack
* [new] 🎉 Scurrius
* [new] 🎉 Tormented Demons
* [new] 🎉 Tombs Of Amascut

Scurrius, and ToA are fun trivial ones to add as the pre-conditions to
trigger these different attack cooldowns are more easily detected with
the current APIs.

In the case the most complex one is definitely the TDs as they involve
their own state tracking which is adjacent to the main plugin, purely to
know if the Demon has triggered.

This TD class has a map of all the known demons and when they trigger
their vulnerability we can detect this via a spot anim, storing it in
the map. Then when we next attack one of these demons, we do some more
checks to tell if the weapon is valid for an attack speed buff. Note
this is still a WIP, right now it only works for 1 demon because I
didn't have time to investigate how the get all the demons currently
interacting with the player. Furthermore for the most part I only really
expect people to kill them solo for long periods anyway so this is the
most important version to get working.

To support some of these new pre-conditions some of the main plugin code
was moved to a ClientUtils package so that any VariableSpeed class can
share those functions.

I'm very open to different names and ways this could be structured, I
figured it was easier to discuss and come up with other structures if we
have a starting point to point at the things we don't like about it.

Finally I added some more exceptions to the spell list.